examples/agno-coaching-ui/backend/requirements.txt
agno>=1.7.8
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.4.0
python-multipart>=0.0.6
websockets>=12.0
python-dotenv>=1.0.0
openai>=1.40.0